Struct hpack::decoder::Decoder

pub struct Decoder<'a> {
    header_table: HeaderTable<'a>,
}

Decodes headers encoded using HPACK.

For now, incremental decoding is not supported, i.e. it is necessary to pass in the entire encoded representation of all headers to the decoder, rather than processing it piece-by-piece.

Fields
header_table: HeaderTable<'a>
Methods

impl<'a> Decoder<'a>

Represents a decoder of HPACK encoded headers. Maintains the state necessary to correctly decode subsequent HPACK blocks.

fn new() -> Decoder<'a>

Creates a new Decoder with all settings set to default values.

fn with_static_table(static_table: StaticTable<'a>) -> Decoder<'a>

Creates a new Decoder with the given slice serving as its static table.

The slice should contain tuples where the tuple coordinates represent the header name and value, respectively.

Note: in order for the final decoded content to match the encoding (according to the standard, at least), this static table must be the one defined in the HPACK spec.

fn set_max_table_size(&mut self, new_max_size: usize)

Sets a new maximum dynamic table size for the decoder.

fn decode(&mut self, buf: &[u8]) -> DecoderResult

Decode the header block found in the given buffer.

The buffer should represent the entire block that should be decoded. For example, in HTTP/2, all continuation frames need to be concatenated to a single buffer before passing them to the decoder.

fn decode_indexed(&self, buf: &[u8]) -> Result<((Vec<u8>, Vec<u8>), usize), DecoderError>

Decodes an indexed header representation.

fn get_from_table(&self, index: usize) -> Result<(&[u8], &[u8]), DecoderError>

Gets the header (name, value) pair with the given index from the table.

In this context, the "table" references the definition of the table where the static table is concatenated with the dynamic table and is 1-indexed.

fn decode_literal(&mut self, buf: &[u8], index: bool) -> Result<((Vec<u8>, Vec<u8>), usize), DecoderError>

Decodes a literal header representation from the given buffer.

Parameters

index: whether or not the decoded value should be indexed (i.e. included in the dynamic table).

fn update_max_dynamic_size(&mut self, buf: &[u8]) -> usize

Handles processing the SizeUpdate HPACK block: updates the maximum size of the underlying dynamic table, possibly causing a number of headers to be evicted from it.

Assumes that the first byte in the given buffer buf is the first octet in the SizeUpdate block.

Returns the number of octets consumed from the given buffer.