Struct hpack::DynamicTable

struct DynamicTable {
    table: VecDeque<(Vec<u8>, Vec<u8>)>,
    size: usize,
    max_size: usize,
}

A struct representing the dynamic table that needs to be maintained by the coder.

The dynamic table contains a number of recently used headers. The size of the table is constrained to a certain number of octets. If on insertion of a new header into the table, the table would exceed the maximum size, headers are evicted in a FIFO fashion until there is enough room for the new header to be inserted. (Therefore, it is possible that though all elements end up being evicted, there is still not enough space for the new header: when the size of this individual header exceeds the maximum size of the table.)

The current size of the table is calculated, based on the IETF definition, as the sum of sizes of each header stored within the table, where the size of an individual header is len_in_octets(header_name) + len_in_octets(header_value) + 32.

Note: the maximum size of the dynamic table does not have to be equal to the maximum header table size as defined by a "higher level" protocol (such as the SETTINGS_HEADER_TABLE_SIZE setting in HTTP/2), since HPACK can choose to modify the dynamic table size on the fly (as long as it keeps it below the maximum value set by the protocol). So, the DynamicTable only cares about the maximum size as set by the HPACK {en,de}coder and lets it worry about making certain that the changes are valid according to the (current) constraints of the protocol.

Fields

table

size

max_size

Methods

impl DynamicTable

fn new() -> DynamicTable

Creates a new empty dynamic table with a default size.

fn with_size(max_size: usize) -> DynamicTable

Creates a new empty dynamic table with the given maximum size.

fn get_size(&self) -> usize

Returns the current size of the table in octets, as defined by the IETF HPACK spec.

fn iter(&self) -> DynamicTableIter

Returns an Iterator through the headers stored in the DynamicTable.

The iterator will yield elements of type (&[u8], &[u8]), corresponding to a single header name and value. The name and value slices are borrowed from their representations in the DynamicTable internal implementation, which means that it is possible only to iterate through the headers, not mutate them.

fn set_max_table_size(&mut self, new_max_size: usize)

Sets the new maximum table size.

If the current size of the table is larger than the new maximum size, existing headers are evicted in a FIFO fashion until the size drops below the new maximum.

fn get_max_table_size(&self) -> usize

Returns the maximum size of the table in octets.

fn add_header(&mut self, name: Vec<u8>, value: Vec<u8>)

Add a new header to the dynamic table.

The table automatically gets resized, if necessary.

Do note that, under the HPACK rules, it is possible the given header is not found in the dynamic table after this operation finishes, in case the total size of the given header exceeds the maximum size of the dynamic table.

fn consolidate_table(&mut self)

Consolidates the table entries so that the table size is below the maximum allowed size, by evicting headers from the table in a FIFO fashion.

fn len(&self) -> usize

Returns the number of headers in the dynamic table.

This is different than the size of the dynamic table.

fn to_vec(&self) -> Vec<(Vec<u8>, Vec<u8>)>

Converts the current state of the table to a Vec

fn get(&self, index: usize) -> Option<&(Vec<u8>, Vec<u8>)>

Returns a reference to the header at the given index, if found in the dynamic table.