Struct hpack::HeaderTable

struct HeaderTable<'a> {
    static_table: StaticTable<'a>,
    dynamic_table: DynamicTable,
}

The struct represents the header table obtained by merging the static and dynamic tables into a single index address space, as described in section 2.3.3. of the HPACK spec.

Fields
static_table: StaticTable<'a> dynamic_table: DynamicTable
Methods

impl<'a> HeaderTable<'a>

fn with_static_table(static_table: StaticTable<'a>) -> HeaderTable<'a>

Creates a new header table where the static part is initialized with the given static table.

fn iter(&'a self) -> HeaderTableIter<'a>

Returns an iterator through all headers stored in the header table, i.e. it includes both the ones found in the static table and the dynamic table, in the order of their indices in the single address space (first the headers in the static table, followed by headers in the dynamic table).

The type yielded by the iterator is (&[u8], &[u8]), where the tuple corresponds to the header name, value pairs in the described order.

fn add_header(&mut self, name: Vec<u8>, value: Vec<u8>)

Adds the given header to the table. Of course, this means that the new header is added to the dynamic part of the table.

If the size of the new header is larger than the current maximum table size of the dynamic table, the effect will be that the dynamic table gets emptied and the new header does not get inserted into it.

fn get_from_table(&self, index: usize) -> Option<(&[u8], &[u8])>

Returns a reference to the header (a (name, value) pair) with the given index in the table.

The table is 1-indexed and constructed in such a way that the first entries belong to the static table, followed by entries in the dynamic table. They are merged into a single index address space, though.

This is according to the HPACK spec, section 2.3.3.

fn find_header(&self, header: (&[u8], &[u8])) -> Option<(usize, bool)>

Finds the given header in the header table. Tries to match both the header name and value to one of the headers in the table. If no such header exists, then falls back to returning one that matched only the name.

Returns

An Option, where Some corresponds to a tuple representing the index of the header in the header tables (the 1-based index that HPACK uses) and a bool indicating whether the value of the header also matched.