Struct hpack::decoder::Decoder

pub struct Decoder<'a> {
    // some fields omitted
}

Decodes headers encoded using HPACK.

For now, incremental decoding is not supported, i.e. it is necessary to pass in the entire encoded representation of all headers to the decoder, rather than processing it piece-by-piece.

Methods

impl<'a> Decoder<'a>

Represents a decoder of HPACK encoded headers. Maintains the state necessary to correctly decode subsequent HPACK blocks.

fn new() -> Decoder<'a>

Creates a new Decoder with all settings set to default values.

fn set_max_table_size(&mut self, new_max_size: usize)

Sets a new maximum dynamic table size for the decoder.

fn decode(&mut self, buf: &[u8]) -> DecoderResult

Decode the header block found in the given buffer.

The buffer should represent the entire block that should be decoded. For example, in HTTP/2, all continuation frames need to be concatenated to a single buffer before passing them to the decoder.