Module hpack::huffman

A module exposing utilities for encoding and decoding Huffman-coded octet strings, under the Huffman code defined by HPACK. (HPACK-draft-10, Appendix B)

Structs

HuffmanDecoder
A simple implementation of a Huffman code decoder.

Enums

HuffmanDecoderError
Represents the error variants that the HuffmanDecoder can return.

Type Definitions

HuffmanDecoderResult
The type that represents the result of the decode method of the HuffmanDecoder.